Deleting have to be with negative -del

+++ nestkernel/archiving_node.h
@@ -260,8 +260,8 @@ private:
   // Time of the last update of the Calcium concentration in ms
   double Ca_t_;
 
-  // Max number of deleted synapses per update step dependent on the firing rate 
-  int max_delete_z;
+  // Max number of deleted synapses per update step dependent on the firing rate (precentage)
+  double max_delete_z;
 
   // constant deleted synaptic elements per update step
   int const_z_deletion;
